Se implemento (reorganizo) capa Domain e Infrastructure siguiendo estructura de folders
