+++ src/P2/SalesAnalysisPlatform/SalesAnalysisPlatform.Domain/SalesAnalysisPlatform.Domain.csproj
@@ -6,4 +6,8 @@
     <Nullable>enable</Nullable>
   </PropertyGroup>
 
+  <ItemGroup>
+    <Folder Include="Repositories\" />
+  </ItemGroup>
+
 </Project>
