+++ src/P2/SalesAnalysisPlatform/SalesAnalysisPlatform.Infrastructure/Core/BaseRepository.cs
@@ -1,0 +1,12 @@
+﻿using System;
+using System.Collections.Generic;
+using System.Linq;
+using System.Text;
+using System.Threading.Tasks;
+
+namespace SalesAnalysisPlatform.Infrastructure.Core
+{
+    public abstract class BaseRepository<TEntity> where TEntity : class
+    {
+    }
+}
