+++ src/P2/SalesAnalysisPlatform/SalesAnalysisPlatform.Infrastructure/IOC/DependencyInjection.cs
@@ -3,6 +3,8 @@ using Microsoft.Extensions.DependencyInjection;
 using SalesAnalysisPlatform.Domain.Repositories;
 using SalesAnalysisPlatform.Infrastructure.Context;
 using SalesAnalysisPlatform.Infrastructure.Repositories;
+using SalesAnalysisPlatform.Application.Interfaces;
+using SalesAnalysisPlatform.Application.Services;
 
 namespace SalesAnalysisPlatform.Infrastructure.IOC
 {
@@ -14,6 +16,7 @@ namespace SalesAnalysisPlatform.Infrastructure.IOC
                 options.UseOracle(connectionString));
 
             services.AddScoped<ISaleRepository, SaleRepository>();
+            services.AddScoped<ISaleService, SaleService>();
 
             return services;
         }
