+++ src/P2/SalesAnalysisPlatform/SalesAnalysisPlatform.Infrastructure/Models/SaleModel.cs
@@ -1,0 +1,17 @@
+﻿using System;
+using System.Collections.Generic;
+using System.Linq;
+using System.Text;
+using System.Threading.Tasks;
+
+namespace SalesAnalysisPlatform.Infrastructure.Models
+{
+    internal class SaleModel
+    {
+        public int Id { get; set; }
+        public string ProductName { get; set; }
+        public decimal Price { get; set; }
+        public int Quantity { get; set; }
+        public DateTime SaleDate { get; set; }
+    }
+}
