+++ src/P2/SalesAnalysisPlatform/SalesAnalysisPlatform.Infrastructure/SalesAnalysisPlatform.Infrastructure.csproj
@@ -17,6 +17,7 @@
   </ItemGroup>
 
   <ItemGroup>
+    <ProjectReference Include="..\SalesAnalysisPlatform.Application\SalesAnalysisPlatform.Application.csproj" />
     <ProjectReference Include="..\SalesAnalysisPlatform.Domain\SalesAnalysisPlatform.Domain.csproj" />
   </ItemGroup>
